change auth checks
