+++ frontend/src/components/CourseModal/CourseModal.tsx
@@ -16,6 +16,7 @@ import { useFerry } from '../../contexts/ferryContext';
 import type { Season, Crn, Listing } from '../../utilities/common';
 import { CUR_YEAR } from '../../config';
 import CourseConflictIcon from '../Search/CourseConflictIcon';
+import { checkAuth } from '../../utilities/api';
 
 const extraInfoMap: { [info in Listing['extra_info']]: string } = {
   ACTIVE: 'ACTIVE',
@@ -99,6 +100,17 @@ const CourseModalEvaluations = suspended(
 function CourseModal() {
   const [searchParams, setSearchParams] = useSearchParams();
   const { requestSeasons, courses } = useFerry();
+  const [isAuthenticated, setIsAuthenticated] = useState<boolean | undefined>(
+    undefined,
+  );
+  useEffect(() => {
+    // Better way to do auth?
+    const init = async () => {
+      const authStatus = await checkAuth();
+      setIsAuthenticated(authStatus);
+    };
+    init();
+  }, []);
 
   const [view, setView] = useState<'overview' | 'evals'>('overview');
   // Stack for listings that the user has viewed
@@ -193,7 +205,10 @@ function CourseModal() {
                   {
                     label: 'Evaluations',
                     value: 'evals',
-                    hidden: CUR_YEAR.includes(listing.season_code),
+                    // don't show eval tab if it's current year or no auth
+                    hidden:
+                      CUR_YEAR.includes(listing.season_code) ||
+                      !isAuthenticated,
                   },
                 ]}
                 onSelectTab={setView}
@@ -214,6 +229,7 @@ function CourseModal() {
         {view === 'overview' ? (
           // Show overview data
           <CourseModalOverview
+            isAuthenticated={isAuthenticated}
             gotoCourse={(l) => {
               setView('evals');
               if (
